backend/requirements.txt
# =============================================================================
# PCONFIG RECOMMENDER API - PYTHON DEPENDENCIES
# =============================================================================
# Production dependencies for the FastAPI-based recommendation system
# 
# Core Framework and Web Server
fastapi==0.104.1                    # Modern, fast web framework for building APIs
uvicorn[standard]==0.24.0           # Lightning-fast ASGI server
pydantic==2.5.0                     # Data validation and settings management
pydantic-settings==2.1.0            # Settings management with Pydantic

# Database Drivers and ORM
neo4j==5.14.1                       # Neo4j graph database driver
asyncpg==0.29.0                     # Async PostgreSQL adapter for Python
psycopg2-binary==2.9.9              # PostgreSQL adapter (binary distribution)

# Authentication and Security  
pyjwt==2.8.0                         # JWT token creation and verification
python-jose[cryptography]==3.3.0    # Alternative JWT implementation
passlib[bcrypt]==1.7.4               # Password hashing library
bcrypt==4.1.2                        # Password hashing (explicit dependency)
python-multipart==0.0.6             # Multipart form data parsing
cryptography>=42.0.0                # Cryptographic recipes and primitives

# HTTP Client and Requests
httpx==0.25.2                       # Async HTTP client
aiohttp==3.9.1                      # Async HTTP client/server framework
requests==2.31.0                    # HTTP library for synchronous requests

# Data Processing and Utilities
python-dateutil==2.8.2              # Extensions to the standard datetime module
pytz==2023.3                        # World timezone definitions
typing-extensions==4.8.0            # Backported and experimental type hints

# Environment and Configuration
python-dotenv==1.0.0                # Load environment variables from .env file

# Logging and Monitoring
structlog==23.2.0                   # Structured logging
prometheus-client==0.19.0           # Prometheus metrics client
loguru==0.7.2                       # Enhanced logging

# Caching (Optional - Redis support)
redis==5.0.1                        # Redis client
aioredis==2.0.1                     # Async Redis client

# AI/ML Libraries (LangGraph integration)
openai==1.3.7                       # OpenAI API client
langchain>=0.1.0                    # LangChain framework for AI applications
langgraph>=0.0.62                   # LangGraph for multi-agent workflows
langsmith>=0.1.0                    # LangSmith for AI workflow tracing

# Development and Testing (commented out for production)
# pytest==7.4.3                     # Testing framework
# pytest-asyncio==0.21.1            # Async testing support
# httpx==0.25.2                     # HTTP client for testing
# faker==20.1.0                     # Generate fake data for testing

# Data Validation and Serialization
email-validator==2.1.0              # Email address validation
phonenumbers==8.13.26               # Phone number parsing and validation

# Async Utilities
asyncio==3.4.3                      # Asynchronous I/O support
aiofiles==23.2.1                    # Async file operations

# Performance and Optimization
orjson==3.9.10                      # Fast JSON serialization
ujson==5.8.0                        # Ultra fast JSON encoder/decoder

# =============================================================================
# VERSION CONSTRAINTS EXPLANATION
# =============================================================================
# - FastAPI 0.104.1: Latest stable version with async support
# - Neo4j 5.14.1: Compatible with Neo4j 5.x database versions
# - asyncpg 0.29.0: High-performance async PostgreSQL driver
# - pydantic 2.5.0: V2 with improved performance and validation
# - python-jose: JWT handling with cryptographic security
# - All versions pinned for reproducible deployments
#
# =============================================================================
# INSTALLATION INSTRUCTIONS
# =============================================================================
# 
# 1. Create virtual environment:
#    python -m venv venv
#    source venv/bin/activate  # On Windows: venv\Scripts\activate
#
# 2. Upgrade pip and install dependencies:
#    pip install --upgrade pip
#    pip install -r requirements.txt
#
# 3. Verify installation:
#    python -c "import fastapi, neo4j, asyncpg; print('All dependencies installed successfully')"
#
# =============================================================================
# OPTIONAL DEPENDENCIES
# =============================================================================
# 
# For development and testing, install additional packages:
# pip install pytest pytest-asyncio httpx faker black isort mypy
#
# For production monitoring:
# pip install sentry-sdk prometheus-client
#
# For enhanced caching:
# pip install redis aioredis
#